remove cast

+++ Oui Uninteractive/Oui Uninteractive/Include/PlayerShooting.h
@@ -80,7 +80,7 @@ public:
 				}
 
 				// Set bullet spawn point
-				bulletSpawnAngle = atan2(static_cast<float>(mousePos.y) - playerPos.y, static_cast<float>(mousePos.x) - playerPos.x);
+				bulletSpawnAngle = atan2(mousePos.y - playerPos.y, mousePos.x - playerPos.x);
 				bulletSpawnOffset = (playerScaleX >= playerScaleY) ? playerScaleX : playerScaleY;
 				bulletSpawnOffset *= 1.1f;
 				bulletSpawnPos.x = playerPos.x + bulletSpawnOffset * cos(bulletSpawnAngle);
